reorderd settings only used in tests
