+++ tests/test_zip_endpoint.py
@@ -16,7 +16,7 @@ from auth_mail.authentication import (
 )
 from auth_mail.generate_token import create_authz_token
 from main.utils_azure import get_blob_from_storage_account, get_queue_client
-from tests.test_iiif import (
+from tests.test_settings import (
     IMAGE_BINARY_DATA,
     PRE_WABO_IMG_URL_BASE,
     PRE_WABO_IMG_URL_WITH_EXTRA_REFERENCE,
